+++ wezterm-gui/src/termwindow/render/paint.rs
@@ -251,6 +251,7 @@ impl crate::TermWindow {
             .context("filled_rectangle for window background")?;
         }
 
+        let mut float_layers = float_layer.quad_allocator();
         for pos in panes {
             if pos.is_active {
                 self.update_text_cursor(&pos);
@@ -259,14 +260,12 @@ impl crate::TermWindow {
                     mux::Mux::get().record_focus_for_current_identity(pos.pane.pane_id());
                 }
             }
-            self.paint_pane(&pos, &mut layers).context("paint_pane")?;
-        }
-
-        if let Some(float_pane) = self.get_float_pane_to_render() {
-            let mut float_layers = float_layer.quad_allocator();
-
-            self.paint_pane(&float_pane, &mut float_layers).context("paint_pane")?;
-            self.paint_float_border(float_pane, &mut float_layers).context("paint_float_border")?;
+            if pos.is_float {
+                self.paint_pane(&pos, &mut float_layers).context("paint_pane")?;
+                self.paint_float_border(pos, &mut float_layers).context("paint_float_border")?;
+            } else {
+                self.paint_pane(&pos, &mut layers).context("paint_pane")?;
+            }
         }
 
         if let Some(pane) = self.get_active_pane_or_overlay() {
